+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Goal `upgrade-bidy` to remove the package and re-install it.
+- Goal `upgrade-bydi` to remove the package and re-install it.
 - Goal `test-tagged` that allows passing `TEST_TAG` to only run tests
   with that tag. The reverse goal `test-not-tagged` was also added.
 - Goal `test-selector` that allows passing `TEST_SELECTOR_STRING` to